+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [0.0.12-beta] - 2024-09-08
+
+### Changed
+
+- Added wizard for `bbai init`; will re-use existing config values for defaults if present
+- Changed git to be optional
+- Split config handling into global and project
+- Improved handling for API control (port number)
+
+
 ## [0.0.11-alpha] - 2024-09-06
 
 ### Changed